Разбил файл интерфейсов сцены
