+++ ViewModels/Interfaces/IFigure.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using Paint2.ViewModels.Utils;
+
+namespace Paint2.ViewModels.Interfaces
+{
+    public interface IFigure : ISceneObject
+    {
+        bool IsInternal(Point p, double eps);
+        IFigure Intersect(IFigure other);
+        IFigure Union(IFigure other);
+        IFigure Subtract(IFigure other);
+        void SetParameters(IDictionary<string, double> doubleParams, IDictionary<string, Point> pointParams);
+    }
+}
